refactor: move auth config to dedicated file and fix type errors
