+++ apps/web/src/app/api/github/check-access/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../../auth/[...nextauth]/route';
+import { authOptions } from '@/lib/auth';
 
 interface GitHubUrlParts {
   owner: string;
